# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an API that allows users to upload and save recipes. It also allows for them to add the ingredients they currently have on hand and return any suggested recipes for those ingredients.",
    "version": "1.0.0",
    "title": "Recipe Wizard",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "jakobtaze@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8888",
  "basePath": "/",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/recipes": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "recipes"
        ],
        "summary": "Get all recipes",
        "description": "",
        "operationId": "getRecipes",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RecipesGetResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "recipes"
        ],
        "summary": "Create a new recipe",
        "description": "",
        "operationId": "postRecipe",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Recipe object to be created.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RecipePostResponse"
            }
          }
        }
      }
    },
    "/recipes/{recipeId}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "recipes"
        ],
        "summary": "Find recipe by ID",
        "description": "Returns a single recipe",
        "operationId": "getRecipeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "description": "ID of recipe to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RecipeGetResponse"
            }
          }
        }
      }
    },
...
    "/users/{userId}/recipes": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Get user's saved recipes",
        "description": "",
        "operationId": "getUserRecipes",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SavedRecipes"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Save an recipe",
        "description": "",
        "operationId": "saveRec",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Recipe to save",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "recipeId": {
                  "type": "number"
                }
              }
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/users/{userId}/recipes/{savedRecipeId}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Get one of a user's saved recipes",
        "description": "",
        "operationId": "getUserRecipe",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "savedRecipeId",
            "in": "path",
            "description": "ID of saved recipe to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SavedRecipe"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Delete one of a user's saved recipes",
        "description": "",
        "operationId": "deleteUserRecipe",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "savedRecipeId",
            "in": "path",
            "description": "ID of saved recipe to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Recipe": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "RecipePostResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "entry": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            }
          }
        }
      }
    },
    "RecipesGetResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "recipes": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "RecipeGetResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "recipe": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "integer"
            },
            "Title": {
              "type": "string"
            },
            "Description": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "SavedRecipes": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/SavedRecipe"
      }
    },
    "SavedRecipe": {
      "type": "object",
      "properties": {
        "Id": {
          "type": "number"
        },
        "RecipeId": {
          "type": "number"
        },
        "UserId": {
          "type": "number"
        }
      }
    }
  }
}